CLI: added --all option to measures list
