+++ Tinkwell.Cli/Commands/Measures/Reporter.cs
@@ -1,5 +1,4 @@
 using Spectre.Console;
-using System.Collections.Generic;
 using Tinkwell.Measures;
 using Tinkwell.Services;
 
@@ -7,20 +6,24 @@ namespace Tinkwell.Cli.Commands.Measures;
 
 static class Reporter
 {
-    public static Task PrintToConsoleAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues, bool verbose)
+    public static Task PrintToConsoleAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues, bool verbose, bool all)
     {
         if (verbose)
-            return PrintVerboseAsync(measures, withValues);
+            return PrintVerboseAsync(measures, withValues, all);
         else
-            return PrintCompactAsync(measures, withValues);
+            return PrintCompactAsync(measures, withValues, all);
     }
 
-    private static async Task PrintCompactAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues)
+    private static async Task PrintCompactAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues, bool all)
     {
         var table = new SimpleTable("Name", "Type", "Unit", withValues ? "Value" : null);
         await foreach (var entry in measures)
         {
             var definition = withValues ? entry.Measure.Definition : entry.Info.Definition;
+
+            if (!all && definition.HasAttribute(StoreMeasureAttributes.System))
+                continue;
+
             table.AddUnescapedRow(
                 $"[cyan]{definition.Name.EscapeMarkup()}[/]",
                 definition.QuantityType.EscapeMarkup(),
@@ -32,7 +35,7 @@ static class Reporter
         AnsiConsole.Write(table.ToSpectreTable());
     }
 
-    private static async Task PrintVerboseAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues)
+    private static async Task PrintVerboseAsync(IAsyncEnumerable<SearchResponse> measures, bool withValues, bool all)
     {
         var table = new PropertyValuesTable();
         int index = 1;
@@ -40,6 +43,9 @@ static class Reporter
         {
             var definition = withValues ? entry.Measure.Definition : entry.Info.Definition;
             var metadata = withValues ? entry.Measure.Metadata: entry.Info.Metadata;
+
+            if (!all && definition.HasAttribute(StoreMeasureAttributes.System))
+                continue;
 
             table
                 .AddNote($"Mesure {index++}")
