+++ Tinkwell.Services.Proto/Measures/StoreExtensions.cs
@@ -35,39 +35,7 @@ public static class StoreExtensions
 
     public static StoreFacade AsFacade(this Store.StoreClient client)
         => new StoreFacade(client);
-}
 
-public sealed class StoreFacade
-{
-    internal StoreFacade(Store.StoreClient client)
-    {
-        _client = client;
-    }
-
-    public async Task WriteQuantityAsync(string name, double value, CancellationToken cancellationToken)
-    {
-        var request = new StoreUpdateRequest();
-        request.Name = name;
-        request.Value = value.ToStoreValue();
-        await _client.UpdateAsync(request, cancellationToken: cancellationToken);
-    }
-
-    public async Task WriteQuantityAsync(string name, string value, CancellationToken cancellationToken)
-    {
-        var request = new StoreSetMeasureValueRequest();
-        request.Name = name;
-        request.Timestamp = Timestamp.FromDateTime(DateTime.UtcNow);
-        request.ValueString = value;
-        await _client.SetMeasureValueAsync(request, cancellationToken: cancellationToken);
-    }
-
-    public async Task WriteStringAsync(string name, string value, CancellationToken cancellationToken)
-    {
-        var request = new StoreUpdateRequest();
-        request.Name = name;
-        request.Value = value.ToStoreValue();
-        await _client.UpdateAsync(request, cancellationToken: cancellationToken);
-    }
-
-    private readonly Store.StoreClient _client;
+    public static bool HasAttribute(this StoreDefinition definition, StoreMeasureAttributes attributes)
+        => (definition.Attributes & (int)attributes) != 0;
 }
